# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Budget Tracker API",
        "description": "API for managing personal budget transactions",
        "version": "1.0.0"
    },
    "paths": {
        "/transactions": {
            "post": {
                "description": "Add a new transaction to the budget",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "body",
                        "description": "Transaction to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction created"
                    }
                }
            },
            "get": {
                "description": "Get a list of all transactions",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "A list of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Transaction"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "put": {
                "description": "Update an existing transaction",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "transaction",
                        "in": "body",
                        "description": "Transaction data to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction updated"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Transaction": {
            "type": "object",
            "required": ["title", "amount", "date"],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "readOnly": true
                },
                "title": {
                    "type": "string",
                    "example": "Grocery Shopping"
                },
                "amount": {
                    "type": "number",
                    "format": "float",
                    "example": 150.75
                },
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2023-01-23"
                }
            }
        }
    }
}